--- app/forms/decidim/admin/galdakao_zone_form.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-module Decidim
-  module Admin
-    class GaldakaoZoneForm < Form
-      mimic :galdakao_zone
-
-      attribute :name, String
-
-      validates :name, presence: true
-    end
-  end
-end
