--- app/views/layouts/decidim/admin/_header.html.erb
@@ -1,9 +1,0 @@
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<%= csrf_meta_tags %>
-<%= legacy_favicon %>
-<%= favicon %>
-<%= apple_favicon %>
-<%= append_stylesheet_pack_tag "decidim_admin_overrides", media: "all" %>
-<%= stylesheet_pack_tag "decidim_core", "decidim_admin", media: "all" %>
-<%= javascript_pack_tag "decidim_core", "decidim_admin", defer: false %>
-<%= snippets.display(:head) %>
