+++ config/application.rb
@@ -6,12 +6,15 @@ require "decidim/rails"
 require "action_cable/engine"
 # require "action_mailbox/engine"
 # require "action_text/engine"
-require_relative "../lib/decidim/galdakao_census/admin_engine"
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 module DecidimApp
   class Application < Rails::Application
     config.load_defaults 7.2
+    # Settings in config/environments/* take precedence over those specified here.
+    # Application configuration can go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded after loading
+    # the framework and any gems in your application.
   end
 end
