--- config/initializers/galdakao_census.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-Rails.application.config.to_prepare do
-  GaldakaoStreet
-  GaldakaoWebservice
-  CensusActionAuthorizer
-end
