+++ gems/decidim-galdakao_census/app/forms/decidim/galdakao_census/admin/galdakao_zone_form.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+module Decidim
+  module GaldakaoCensus
+    module Admin
+      class GaldakaoZoneForm < Form
+        mimic :galdakao_zone
+        attribute :name, String
+        validates :name, presence: true
+      end
+    end
+  end
+end
