+++ gems/decidim-galdakao_census/app/overrides/decidim/admin/authorization_workflows/add_management_td.html.erb.deface
@@ -1,0 +1,9 @@
+<!-- insert_after "tbody tr" -->
+<td class="!text-left">
+  <% if workflow.key == "census_authorization_handler" %>
+    <div class="flex gap-2">
+      <%= link_to t("decidim.admin.galdakao.streets.title"), decidim_admin_galdakao_census.galdakao_index_path, class: "button button__sm button__secondary" %>
+      <%= link_to t("decidim.admin.galdakao.zones.index.title"), decidim_admin_galdakao_census.galdakao_zones_path, class: "button button__sm button__secondary" %>
+    </div>
+  <% end %>
+</td>
