+++ gems/decidim-galdakao_census/app/overrides/decidim/admin/authorization_workflows/add_management_th.html.erb.deface
@@ -1,0 +1,2 @@
+<!-- insert_after "thead tr" -->
+<th class="!text-left"><%= t("decidim.admin.galdakao.authorization_workflows.management") %></th>
