+++ gems/decidim-galdakao_census/app/packs/src/decidim/admin/application.js
