+++ gems/decidim-galdakao_census/app/views/decidim/galdakao_census/admin/galdakao/index.html.erb
