+++ gems/decidim-galdakao_census/app/views/decidim/galdakao_census/admin/zones/show.html.erb
